Copy token to clipboard automatically on token click
This mimicks the behavior of Google Authenticator and simplifies the UI.

Also, remove the title of the clipboard data so as not leak metadata
about the data in the clipboard. When the title was specified, third
parties could tell that the data in the clipboard was token data.
Because clipboard copying was a manual process, this was not so
problematic. However, now that clipboard copying is on by default,
third parties can snoop this data more easily. So we should at least
avoid broadcasting that the data in the clipboard is a token code.

Fixes #28