Don't reset token state on orientation change
There were three options to do this:
1. Marshall/unmarshall state on activity reload
2. Don't reload activity on orientation change
3. Use a persistent fragment

None of these were particularly great options. However, after several
attempts, option 2 came to be the easiest.

One problem is that, even without reloading the activity, the adapter
would reload the views, resetting the state. To solve this problem,
we now cache state in memory in the adapter. This also has an advantage
that tokens which scroll off the screen do not have their state reset,
which is a more natural UI.

Fixes #22