Fix app crash when scanning empty QR
Once the QR code has been decoded, the app encodes the data again to
display the QR code as a bitmap. However, ZXing refuses to encode an
empty string and throws an uncatched IllegalArgumentException in this
case.

To avoid the issue, validate the data before passing them to ZXing and
display a blank bitmap when invalid. By the way, also display a blank
image when ZXing returns a documented WriterException when trying to
encode the data.

Fixes #412