Remove deprecated android.defaults.buildfeatures.buildconfig
Following warning appears when compiling the project with Gradle:
The option setting 'android.defaults.buildfeatures.buildconfig=true' is deprecated.
The current default is 'false'.
It will be removed in version 10.0 of the Android Gradle plugin.
To keep using this feature, add the following to your module-level build.gradle files:
    android.buildFeatures.buildConfig = true
or from Android Studio, click: `Refactor` > `Migrate BuildConfig to Gradle Build Files`.

And as the BuildConfig does not seem to be used by the project, moving
to android.buildFeatures.buildConfig = true is not even necessary.